items

+++ src/app/routers/router-base.ts
@@ -1,6 +1,8 @@
 import * as restify from 'restify';
 import { IRouter } from '../irouter';
 import { IEntity, IFilter, IChange, IManager, DuplicateEntityError } from '../framework';
+import { IUser } from '../framework/user';
+import { IPermission } from '../framework/security';
 import { IRequest } from '../irequest';
 import { IResponse } from '../iresponse';
 import { PermissionHelper } from '../security';
@@ -50,7 +52,7 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
           return response.send(new restify.UnauthorizedError());
 
         if (permissions && permissions.length !== 0)
-          if (!permissions.every(permission => PermissionHelper.hasPermission(request.permissions, permission)))
+          if (!permissions.every(permission => PermissionHelper.hasPermission(request.permissions, permission, {})))
             return response.send(new restify.ForbiddenError());
       }
 
@@ -60,10 +62,11 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
 
   protected async getEntities(request: IRequest, response: IResponse) {
     const params = new Params(request);
-    const filter = await this.filterFromParams(params);
+    const filter = await this.filterFromParams(params, request);
     const entities = await this.manager.getAll(filter);
 
-    const data = entities.map(entity => this.entityToModel(entity));
+    const accessibleEntities = entities.filter(entity => this.checkEntityAccess(entity, 'read', request.user, request.permissions));
+    const data = accessibleEntities.map(entity => this.entityToModel(entity));
 
     response.send({
       data: data,
@@ -77,6 +80,9 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
     if (!entity)
       return response.send(new restify.NotFoundError());
 
+    if (!this.checkEntityAccess(entity, 'read', request.user, request.permissions))
+      return response.send(new restify.ForbiddenError());
+
     const data = this.entityToModel(entity);
 
     response.send({
@@ -86,12 +92,15 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
 
   protected async postEntity(request: IRequest, response: IResponse) {
     const params = new Params(request);
-    const entity = await this.entityFromParams(params);
+    const entity = await this.entityFromParams(params, request);
 
     const validationError = this.manager.validateEntity(entity);
 
     if (validationError)
       return response.send(new restify.UnprocessableEntityError(validationError.message));
+
+    if (!this.checkEntityAccess(entity, 'write', request.user, request.permissions))
+      return response.send(new restify.ForbiddenError());
 
     try {
       const insertedEntity = await this.manager.insert(entity);
@@ -116,12 +125,18 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
     if (!entity)
       return response.send(new restify.NotFoundError());
 
-    const change = await this.changeFromParams(params);
+    if (!this.checkEntityAccess(entity, 'write', request.user, request.permissions))
+      return response.send(new restify.ForbiddenError());
+
+    const change = await this.changeFromParams(params, request);
 
     const validationError = this.manager.validateChange(change);
 
     if (validationError)
       return response.send(new restify.UnprocessableEntityError(validationError.message));
+
+    if (!this.checkChangeAccess(change, request.user, request.permissions))
+      return response.send(new restify.ForbiddenError());
 
     try {
       const updatedEntity = await this.manager.update(entity.id, change);
@@ -146,21 +161,32 @@ export abstract class RouterBase<TEntity extends IEntity, TFilter extends IFilte
     if (!entity)
       return response.send(new restify.NotFoundError());
 
+    if (!this.checkEntityAccess(entity, 'write', request.user, request.permissions))
+      return response.send(new restify.ForbiddenError());
+
     await this.manager.delete(entity.id);
 
     response.send(200);
   }
 
-  protected filterFromParams(params: IParams) {
+  protected filterFromParams(params: IParams, request: IRequest) {
     return Promise.resolve({} as TFilter);
   }
 
-  protected entityFromParams(params: IParams) {
+  protected entityFromParams(params: IParams, request: IRequest) {
     return Promise.resolve({} as TEntity);
   }
 
-  protected changeFromParams(params: IParams) {
+  protected changeFromParams(params: IParams, request: IRequest) {
     return Promise.resolve({} as TChange);
+  }
+
+  protected checkEntityAccess(entity: TEntity, access: string, user: IUser, permissions: IPermission[]) {
+    return true;
+  }
+
+  protected checkChangeAccess(change: TChange, user: IUser, permissions: IPermission[]) {
+    return true;
   }
 
   protected entityToModel(entity: TEntity) {
